requirements.txt
# Backend Core
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0.post2

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7  # Για PostgreSQL
alembic==1.12.0

# HTTP Client
httpx==0.24.1
aiohttp==3.8.5

# AI & ML
torch==2.0.1
transformers==4.33.2
peft==0.5.0
bitsandbytes==0.41.1
accelerate==0.23.0

# Utilities
python-telegram-bot==20.5
redis==5.0.0
celery==5.3.4
numpy==1.25.2
pandas==2.1.0
aiocache==0.12.1
scikit-learn==1.3.0
matplotlib==3.7.2

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Development tools
black==23.7.0
isort==5.12.0
flake8==6.1.0

# Monitoring & Logging
prometheus-client==0.17.1
python-json-logger==2.0.7
